Update QR file names
There may be several teams with the same slug, so it's necessary to
tack the unique ID on the filename to make sure we're pulling up the
right team for the right competition.

Also added the qr filename to the Team instances to make it easier to
reference the file.